Refine unitTestThreealNumber tolerance and comments
Reduced the numerical comparison tolerance from 1e-1 to 1e-2 for more precise testing. This is possible because the new evaluation point is at 0.25 instead of 0.5 where the numerical differentiation is apparently more accurate. The main reason to change the evaluation point was to avoid a negative result though - because the square-root function can't work on negative inputs